backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.52",
    "fastapi>=0.104.1",
    "google-genai>=1.63.0",
    "jinja2>=3.1.6",
    "mangum>=0.21.0",
    "networkx>=3.6.1",
    "openai>=2.21.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "pymupdf-layout>=1.27.1",
    "pymupdf4llm>=0.3.4",
    "pypdf>=6.7.1",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.22",
    "qstash>=3.2.0",
    "reportlab>=4.4.10",
    "requests>=2.31.0",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.46",
    "upstash-redis>=1.6.0",
    "uvicorn[standard]>=0.24.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:.*importlib._bootstrap",
    "ignore:.*SwigPyPacked.*:DeprecationWarning",
    "ignore:.*SwigPyObject.*:DeprecationWarning",
    "ignore:.*swigvarlink.*:DeprecationWarning",
]
